bugfix - in callback API deliver timestamp with single digit after dot

+++ alisamqttbridge/src/main/java/by/ibn/alisamqttbridge/resources/CallbackRequest.java
@@ -1,5 +1,7 @@
 package by.ibn.alisamqttbridge.resources;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
@@ -10,7 +12,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class CallbackRequest {
 	
 	@JsonProperty("ts")
-	public Float timestamp;
+	public BigDecimal timestamp;
 	
 	@JsonProperty("payload")
 	public Payload payload;
